+++ .idea/misc.xml
@@ -19,6 +19,19 @@
         <entry key="../../../../layout/compose-model-1661151719216.xml" value="0.33" />
         <entry key="../../../../layout/compose-model-1661151845206.xml" value="0.75" />
         <entry key="../../../../layout/compose-model-1661173716742.xml" value="0.25" />
+        <entry key="../../../../layout/compose-model-1661273199469.xml" value="0.1" />
+        <entry key="../../../../layout/compose-model-1661273470819.xml" value="0.5" />
+        <entry key="../../../../layout/compose-model-1661273734819.xml" value="0.5" />
+        <entry key="../../../../layout/compose-model-1661274792158.xml" value="0.5" />
+        <entry key="../../../../layout/compose-model-1661277468344.xml" value="0.319047619047619" />
+        <entry key="../../../../layout/compose-model-1661278024516.xml" value="0.31165540540540543" />
+        <entry key="../../../../layout/compose-model-1661278040256.xml" value="0.33" />
+        <entry key="../../../../layout/compose-model-1661278119284.xml" value="0.6314814814814815" />
+        <entry key="../../../../layout/compose-model-1661279385571.xml" value="0.25386996904024767" />
+        <entry key="../../../../layout/compose-model-1661280173031.xml" value="0.25" />
+        <entry key="../../../../layout/compose-model-1661281559240.xml" value="0.33" />
+        <entry key="../../../../layout/compose-model-1661332423815.xml" value="0.12210072895957588" />
+        <entry key="../../../../layout/compose-model-1661332638798.xml" value="0.18" />
         <entry key="app/src/main/res/drawable/ic_pin.xml" value="0.2385" />
         <entry key="app/src/main/res/font/a.xml" value="0.37948717948717947" />
       </map>
