+++ app/src/main/java/com/otaman/weather/ui/MainActivity.kt
@@ -6,6 +6,7 @@ import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.*
 import androidx.compose.ui.Modifier
+import com.otaman.weather.ui.main_screen.MainScreen
 import com.otaman.weather.ui.theme.WeatherTheme
 
 class MainActivity : ComponentActivity() {
@@ -18,7 +19,7 @@ class MainActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colors.background
                 ) {
-                    WeatherApp()
+                    MainScreen()
                 }
             }
         }
